Hybrid Automation FW using Selenium
Hybrid Automation Framework for Canvas Web Application. mainDriver file
calls the configFile() to get the browser & env details. mainDriver
calls the testFunction() which inturn calls the individual tests based
on the test cases from excel file. Based on the size of course list to
be tested the script navigates through all the pages & tests the page.
The result is written to an excel file with expected & actual results.